+++ console/src/axios.ts
@@ -1,0 +1,5 @@
+import axios_module from 'axios'
+
+const axios = axios_module.create({})
+
+export default axios