+++ console/src/components/TOTPView.vue
@@ -1,8 +1,5 @@
 <script setup lang="ts">
-import Fetch from '@/fetcher'
 import { VOtpInput } from 'vuetify/components'
-import useNotifications from '@/stores/notifications'
-import useMessages from '@/stores/messages'
 
 const router = useRouter()
 const notifications = useNotifications()
@@ -15,29 +12,29 @@ onMounted(() => {
   input.value!.focus()
 })
 
-function submitLogin() {
+function verifyTOTP() {
   const busy = messages.addMessage({ text: 'Verifying TOTP...', timeout: 0 })
-  Fetch('/api/v1/verify_totp')
-    .onError((error: any) => {
-      iserror.value = true
-      console.log(error)
-      notifications.post({ text: 'Failed to verify TOTP: ' + error, type: 'error' })
+  axios
+    .post('/api/v1/verify_totp', {
+      totp: totp.value
     })
-    .onStatus(401, () => {
-      console.log('Unauthorized. Redirecting to login page.')
-      router.push('/login')
-    })
-    .onSuccess((data: any) => {
-      if (data.success) {
+    .then((response) => {
+      if (response!.data.success) {
         notifications.post({ text: 'Logged in.', type: 'success' })
         router.push('/agents')
       } else {
-        throw new Error(data.detail)
+        throw new Error(response!.data.detail)
       }
     })
-    .credentials('same-origin')
-    .post({
-      totp: totp.value
+    .catch((error) => {
+      if (error.response?.status == 401) {
+        notifications.post({ text: 'Please log in', type: 'error' })
+        router.push('/login')
+        return
+      }
+      iserror.value = true
+      console.log(error)
+      notifications.post({ text: 'Failed to verify TOTP: ' + error, type: 'error' })
     })
     .finally(() => {
       messages.removeMessage(busy)
@@ -48,13 +45,13 @@ function submitLogin() {
 function verify_totp_when_done() {
   iserror.value = false
   if (totp.value.length === 6) {
-    submitLogin()
+    verifyTOTP()
   }
 }
 </script>
 
 <template>
-  <v-form @submit.prevent="submitLogin">
+  <v-form @submit.prevent="verifyTOTP">
     <v-card max-width="500" title="Verify 2FA" class="mx-auto my-2">
       <v-card-text class="d-flex flex-column ga-4">
         <v-label>Enter the 6-digit code from your authenticator app:</v-label>