+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This is the changelog for RedPepper.
 
 - Migrate console code to TypeScript.
 - Use Vuetify for the console UI.
+- Use `axios` for API requests in the console.
 
 ## [0.0.14]